requirements/dev.txt
Flask==0.10.1
Flask-Admin==1.0.9
Flask-DebugToolbar==0.9.2
Flask-Login==0.2.11
Flask-Migrate==1.3.0
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.10.3
Pillow==2.7.0
Pygments==2.0.1
honcho==0.5.0
misaka==1.0.2
passlib==1.6.2
pyjade==3.0.0
